pyproject.toml
[project]
name = "automation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-kasa>=0.10.2",
    "apscheduler>=3.10.1",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
